teamup/new/copy/requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.4
djangorestframework==3.14.0
PyJWT==1.7.1
PyJWT==2.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
